Data Warehouse/requirements.txt
mysql-connector-python==8.2.0
schedule==1.2.0
